Created a stow and nuke function.  Stow creates a folder and moves all log and txt files into it.  Nuke deletes every log and txt in the folder.  both functions centered around moving the files out of the main logdiver folder to sort of reset the folder for new records.

+++ logdiver.py
@@ -1,19 +1,43 @@
-import os, sys
+import os, sys, shutil
 from warning import warning
 from vmc import vmcCrawl
 
 logName = ""
 logFile = []
 
-def cleaner():
+def cleaner(nukeCheck):
     filesCleaned = 0
-    print("Cleaning all logdiver files...")
-    for file in os.listdir(os.getcwd()):
-        if file.startswith("Warn Log Output - ") or file.startswith(
+    if nukeCheck == 0:
+        for file in os.listdir(os.getcwd()):
+            if file.startswith("Warn Log Output - ") or file.startswith(
                 "Info Log Output - ") or file.startswith("Top Section for ") or file.startswith("test") or file.startswith("No Good - "):
-            os.remove(file)
-            filesCleaned += 1
-    return print("All Clean! Files cleaned: ", filesCleaned)
+                os.remove(file)
+                filesCleaned += 1
+        print("All Clean! Files cleaned: ", filesCleaned)
+    
+    if nukeCheck == 1:
+        nukeCheck = input("Please type YES to finish...")
+        if nukeCheck == "YES":
+            for file in os.listdir(os.getcwd()):
+                if file.endswith(".txt") or file.endswith(".log"):
+                    os.remove(file)
+                    filesCleaned += 1
+            print("Nuclear Option Executed! Files cleaned: ", filesCleaned)
+        else:
+            print("YES in all caps not entered...")
+            print("Nuclear Option Deactivated...")
+            print("{} Files deleted, Exiting...".format(filesCleaned))
+
+def stow(folderName):
+    filesMoved = 0
+    srcPath = os.getcwd()
+    destPath = os.path.join(srcPath, folderName)
+    os.mkdir(destPath)
+    for file in os.listdir(os.getcwd()):
+        if file.endswith(".txt") or file.endswith(".log"):
+            shutil.move(srcPath+file, destPath+file)
+            filesMoved += 1
+    print("{} files stowed away in {}".format(filesMoved, folderName))
 
 def vmcLogCrawl():
     try:
@@ -42,8 +66,18 @@ def logWriter(logFile):
         log.close()
 
 if sys.argv[1] == "clean":
-    cleaner()
+    print("Cleaning all Log diver txt files.")
+    cleaner(0)
     exit()
+        
+if sys.argv[1] == "nuke":
+        print("Nuclear Option Activated...")
+        print("ALL NON-LOGDIVER FILES WILL BE DELETED...")
+        cleaner(1)
+
+if sys.argv[1] == "stow":
+        folderName = input("Name the folder to stow the files in: ")
+        stow(folderName)
     
 if sys.argv[1] == "vmc":
     vmcCrawl()
